Add support for YAML metadata.
If the first line of each .md file is exactly '---', subsequent text up to
but not including a line reading '...' or '---' will be passed to PyYAML
for parsing.  This is consistent with the behavior of the pelican-md-yaml
plugin.  PyYAML need only be installed if this feature is used.

Also, the "plain" key-value metadata syntax is aligned with Python-Markdown.

Also also, hard tabs are now expanded to spaces (this was necessary in
order to align the plain metadata syntax with Python-Markdown) and their
width is controllable in the settings file (new setting `PANDOC_TAB_WIDTH`).