Reimplement metadata extraction and configuration.
We now have Pandoc parse the frontmatter, which means we accept any
frontmatter format it supports, and string values will be processed
as Markdown.

In order to make this work, we need to use “standalone” mode with a
custom template file, which means we need to supply our own “user data
directory” and a defaults file.  It thus becomes simpler to drop
support for PANDOC_ARGS and PANDOC_EXTENSIONS and instruct end-users
to edit the defaults file instead.  To facilitate this, there is now a
utility called `pelican-pandoc-reader-config` that will extract the
embedded user data directory into a Pelican website’s source tree.